fix: 대여 테이블 컬럼 변경으로 CRUD 수정, rent_history CRUD 삭제

+++ bible/src/main/java/com/library/bible/rent/repository/IRentRepository.java
@@ -15,8 +15,8 @@ public interface IRentRepository {
 	List<Rent> selectAllRent();
 	
 	//Rent table CRUD
-	Rent selectRent(int rentId);
+	Rent selectRent(long rentId);
 	void insertRent(Rent rent);
 	void updateRent(Rent rent);
-	int deleteRent(int rentId);
+	int deleteRent(long rentId);
 }
